Add the "imageDownloadURL" field in the "LightboxImage" class that can be used to download a better version of the preview image.
When the "imageDownloadURL" field is set in a "LightboxImage" instance, the library will download the image from this URL (in case the user touches the "Download" button). An activity indicator replaces the download button while the image is being downloaded. When the download finishes, the image will be saved to the "Photos" app, and the "Download" button replaces the activity indicator. In case the "imageDownloadURL" isn't populated, the library will save the preview image (the one that is displayed in the imageView).